# app/api/swagger.yaml
openapi: 3.0.0
info:
  title: MealTime API
  version: 2.0.0
  description: |
    API completa para gerenciamento de alimentação de gatos
    
    **VERSÕES DISPONÍVEIS**:
    - **V2 (Recomendado)**: `/api/v2/*` - Autenticação JWT/Session, respostas padronizadas
    - **V1 (Deprecated)**: `/api/*` - Será removido em 2025-07-28
    
    **AUTENTICAÇÃO V2**:
    - Mobile: JWT via header `Authorization: Bearer <token>`
    - Web: Supabase Session via cookies
  contact:
    name: MealTime Support
    email: support@mealtime.app
servers:
  - url: http://localhost:3000/api/v2
    description: Servidor V2 (desenvolvimento) - RECOMENDADO
  - url: https://mealtime.app/api/v2
    description: Servidor V2 (produção) - RECOMENDADO
  - url: http://localhost:3000/api
    description: Servidor V1 (desenvolvimento) - DEPRECATED
  - url: https://mealtime.app/api
    description: Servidor V1 (produção) - DEPRECATED
...
paths:
...
  /v2/weight-logs:
    get:
      tags: [V2 - Weight]
      summary: 🆕 Listar registros de peso (V2)
      security:
        - bearerAuth: []
      parameters:
        - name: catId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de registros de peso
    
    post:
      tags: [V2 - Weight]
      summary: 🆕 Criar registro de peso (V2)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [catId, weight, date]
              properties:
                catId:
                  type: string
                  format: uuid
                weight:
                  type: number
                date:
                  type: string
                  pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
                notes:
                  type: string
      responses:
        '201':
          description: Registro criado
    
    put:
      tags: [V2 - Weight]
      summary: 🆕 Atualizar registro de peso (V2)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [catId, weight, date]
              properties:
                catId:
                  type: string
                weight:
                  type: number
                date:
                  type: string
      responses:
        '200':
          description: Registro atualizado
    
    delete:
      tags: [V2 - Weight]
      summary: 🆕 Deletar registro de peso (V2)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Registro deletado

  # ==================== V2 - GOALS (RECOMENDADO) ====================
...
  /v2/schedules:
    get:
      tags: [V2 - Schedules]
      summary: 🆕 Listar agendamentos (V2)
      security:
        - bearerAuth: []
      parameters:
        - name: householdId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de agendamentos
    
    post:
      tags: [V2 - Schedules]
      summary: 🆕 Criar agendamento (V2)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [catId, type]
              properties:
                catId:
                  type: string
                  format: uuid
                type:
                  type: string
                  enum: [interval, fixedTime]
                interval:
                  type: integer
                times:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Agendamento criado
...
  /schedules:
    get:
      tags: [Schedules]
      summary: ⚠️ Listar agendamentos (V1 - DEPRECATED)
      description: |
        **DEPRECATED**: Use `/v2/schedules` em vez disso.
        Será removido em 2025-07-28.
      deprecated: true
      responses:
        '200':
          description: Lista de agendamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
    
    post:
      tags: [Schedules]
      summary: Criar agendamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleInput'
      responses:
        '201':
          description: Agendamento criado
...
  /weight-logs:
    get:
      tags: [Weight]
      summary: ⚠️ Histórico de peso (V1 - DEPRECATED)
      description: |
        **DEPRECATED**: Use `/v2/weight-logs` em vez disso.
        Será removido em 2025-07-28.
      deprecated: true
      parameters:
        - name: catId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Histórico de peso do gato
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeightLog'
    
    post:
      tags: [Weight]
      summary: Registrar peso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeightLogInput'
      responses:
        '201':
          description: Peso registrado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtido via POST /api/auth/mobile
...
  schemas:
...
    Schedule:
      type: object
      properties:
        id:
          type: string
        catId:
          type: string
        time:
          type: string
          format: time
        days:
          type: array
          items:
            type: string
        mealType:
          type: string
        active:
          type: boolean
    ScheduleInput:
      type: object
      required: [catId, time, days]
      properties:
        catId:
          type: string
        time:
          type: string
        days:
          type: array
          items:
            type: string
        mealType:
          type: string
        active:
          type: boolean
    WeightLog:
      type: object
      properties:
        id:
          type: string
        catId:
          type: string
        weight:
          type: number
        date:
          type: string
          format: date
        notes:
          type: string
    WeightLogInput:
      type: object
      required: [catId, weight]
      properties:
        catId:
          type: string
        weight:
          type: number
        date:
          type: string
          format: date
        notes:
          type: string
...
security:
  - bearerAuth: []